requirements.txt
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.1.0
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.0.5
cryptography==40.0.1
cssselect2==0.7.0
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==0.5.0
django-object-actions==4.1.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
eight==0.4.2
entrypoints==0.4
fasteners==0.14.1
filelock==3.0.12
fonttools==4.37.4
fs==2.4.12
future==0.16.0
gast==0.5.2
Geraldo @ https://github.com/thiagopena/geraldo/archive/master.zip
greenlet==3.1.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
isort==4.3.21
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
notebook==6.4.8
num2words==0.5.12
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
oscrypto==1.3.0
packaging==21.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==3.12.4
ptyprocess==0.7.0
py==1.10.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
PyJWT==2.3.0
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==17.5.0
pyparsing==2.4.7
pypng==0.20220715.0
pyRFC3339==1.1
pyrsistent==0.18.1
PySIGNFe @ https://github.com/thiagopena/PySIGNFe/archive/master.zip
python-apt==2.4.0+ubuntu4
python-bidi==0.4.2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-decouple==3.1
python-dotenv==0.13.0
pythran==0.10.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.25.1
scipy==1.8.0
SecretStorage==3.3.1
Send2Trash==1.8.1b0
signxml==2.5.2
six==1.13.0
soupsieve==2.3.1
sqlparse==0.4.4
supervisor==4.2.1
svglib==1.5.1
sympy==1.9
systemd-python==234
terminado==0.13.1
testpath==0.5.0
texttable==1.6.4
tinycss2==1.2.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==4.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uritools==4.0.1
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.4.3
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
widgetsnbextension==2.0.0
wrapt==1.11.2
xdg==5
xkit==0.0.0
xlwt==1.3.0
zipp==1.0.0
